+++ crates/bitwarden-auth/src/token_management/mod.rs
@@ -1,10 +1,10 @@
 //! Token renewal module.
 
 mod common;
-mod password_manager;
+mod password_manager_token_handler;
 #[cfg(feature = "secrets")]
-mod secrets_manager;
+mod secrets_manager_token_handler;
 
-pub use password_manager::PasswordManagerTokenHandler;
+pub use password_manager_token_handler::PasswordManagerTokenHandler;
 #[cfg(feature = "secrets")]
-pub use secrets_manager::SecretsManagerTokenHandler;
+pub use secrets_manager_token_handler::SecretsManagerTokenHandler;
