+++ crates/bitwarden-auth/src/token_management/password_manager_token_handler.rs
