+++ crates/bitwarden-auth/src/token_management/secrets_manager_token_handler.rs
